REFACTOR getNodeName onto its own file
