+++ node.go
@@ -1,0 +1,15 @@
+package main
+
+import "os"
+
+// getNodeName retrieves the node name from environment variables
+func getNodeName() string {
+	nodeName := os.Getenv("NODE_NAME")
+	if nodeName == "" {
+		nodeName = os.Getenv("HOSTNAME")
+	}
+	if nodeName == "" {
+		nodeName = "unknown"
+	}
+	return nodeName
+}
